BZ edition, initial commit
Attempted to make the solution buildable for both SN1 and BZ by changing the active configuration; the SN1 configuration will build a DLL, though this DLL has not yet been tested.

The project has been edited to rely on a GameDir.targets file in the solution directory, but GitHub Desktop does not appear to be uploading this file.

BZ version: Turbine components are all 1x1 in inventory. This may be an issue in SMLHelper.